[EA-631] update channels biz service;

+++ app/service/biz/channels/internal/dao/dataobject/channelMessageBox_do.go
@@ -4,7 +4,7 @@ type ChannelMessageBoxDO struct {
 	Id                  int64  `db:"id"`
 	SenderUserId        int64  `db:"sender_user_id"`
 	ChannelId           int64  `db:"channel_id"`
-	ChannelMessageBoxId int64  `db:"channel_message_box_id"`
+	ChannelMessageBoxId int32  `db:"channel_message_box_id"`
 	MessageId           int64  `db:"message_id"`
 	Date                int32  `db:"date"`
 	Deleted             int8   `db:"deleted"`
